fix build for mips loongarch risc-v with simplemath (#6024)

+++ src/layer/mips/binaryop_mips.cpp
@@ -306,12 +306,12 @@ MAKE_FUNCTION(binary_op_mul, x * y, __msa_fmul_w(x, y))
 MAKE_FUNCTION(binary_op_div, x / y, __msa_fdiv_w(x, y))
 MAKE_FUNCTION(binary_op_max, std::max(x, y), __msa_fmax_w(x, y))
 MAKE_FUNCTION(binary_op_min, std::min(x, y), __msa_fmin_w(x, y))
-MAKE_FUNCTION(binary_op_pow, (float)pow(x, y), pow_ps(x, y))
+MAKE_FUNCTION(binary_op_pow, (float)powf(x, y), pow_ps(x, y))
 MAKE_FUNCTION(binary_op_rsub, y - x, __msa_fsub_w(y, x))
 MAKE_FUNCTION(binary_op_rdiv, y / x, __msa_fdiv_w(y, x))
-MAKE_FUNCTION(binary_op_rpow, (float)pow(y, x), pow_ps(y, x))
-MAKE_FUNCTION(binary_op_atan2, (float)atan2(x, y), atan2_ps(x, y))
-MAKE_FUNCTION(binary_op_ratan2, (float)atan2(y, x), atan2_ps(y, x))
+MAKE_FUNCTION(binary_op_rpow, (float)powf(y, x), pow_ps(y, x))
+MAKE_FUNCTION(binary_op_atan2, (float)atan2f(x, y), atan2_ps(x, y))
+MAKE_FUNCTION(binary_op_ratan2, (float)atan2f(y, x), atan2_ps(y, x))
 // *INDENT-ON*
 // clang-format on
